Lesson code quality: round 3 — consistency fixes from review
1. Lesson 13 now owns file-scope variables: a sentence before the
   s13a-shapes bullets, a topic-table row that lists them, and 14-motion's
   opening paragraph softened to recall the idea rather than introduce it.
2. 14-motion said asking for Screen.Width beats typing 640 and then typed
   640 six lines later. s14a-bounce, s14b-gravity, s14c-trails and
   c14-bouncing-ball.solution now test against Screen.Width - radius and
   Screen.Height - radius (arithmetically identical); the prose, the
   gravity try block and both c14 bullets follow. Screen.Height moves to
   the Lesson 14 row of the topic table.
3. s18a-rect-overlap's two bare 5s become playerSpeed.
4. c13-house carries screenWidth, screenHeight and groundY in both starter
   and solution, so the given ground matches s13a-shapes; the statement
   names them.
5. c18-coin's quoted test fence tagged csharp.
6. c19c's statement states 640 and (320, 180) once beside the names.
7. Rewrapped the over-long lines in AUTHORING.md, 19-pong.md, 14-motion.md
   and 15-the-keyboard.md.
8. Three comments that restated rather than explained rewritten or moved
   (c03-sweets, s14b-gravity, s17-rain).
9. Pong's declarations grouped screen+paddles / ball / score, identically
   in all four files; c19c.start is still byte-identical to
   c19b.solution.
10. "roughly a third of the file" -> "about a quarter" in both places.
11. c15-paddle's paddleSpeed is int, matching Pong's.

Build 0 warnings, 221/221, no golden output, frame or frames file touched,
starter gate green.

+++ content/challenges/c15-paddle.start.cs
@@ -4,7 +4,7 @@ int paddleWidth = 100;
 int paddleHeight = 16;
 int paddleY = 330;
 int paddleMaxX = screenWidth - paddleWidth;
-double paddleSpeed = 6;
+int paddleSpeed = 6;
 double x = 270;
 
 void Setup()
